Module Bap_elementary

module Elementary : functor (Theory : Bap_core_theory.Theory.Core) -> sig ... end

Elementary is a library that provides few primitives for approximations of floating point operations via table methods.