Module Bap_elementary.Elementary

Elementary is a library that provides few primitives for approximations of floating point operations via table methods.

Parameters

Theory : Bap_core_theory.Theory.Core

Signature

type 'a t = 'a Bap_knowledge.knowledge

exception Not_a_table

val approximate : rank:int -> reduce:(('a, 's) Bap_core_theory.Theory.format Bap_core_theory.Theory.float -> 'r Bap_core_theory.Theory.bitv) -> extract:(int -> 'd Bap_core_theory.Theory.bitv -> 's Bap_core_theory.Theory.bitv) -> coefs:('r, 'd) Bap_core_theory.Theory.Mem.t Bap_core_theory.Theory.var -> ('a, 's) Bap_core_theory.Theory.format Bap_core_theory.Theory.float -> Bap_core_theory.Theory.rmode -> ('a, 's) Bap_core_theory.Theory.format Bap_core_theory.Theory.float

approximate ~rank ~reduce ~extract ~coefs x rmode returns a function f(x,rmode) that is defined by a polynomial of rank rank, which coefficients are stored in a table coefs.

parameter coefs

is a table, where keys are some points in integer space, in which floating point values could be mapped by reduce function. And values of coefs are some point in integer namespace from which floating point coefficients could be restored by extract function with a respect to a rank of each coefficient.

val tabulate : string -> rank:int -> size:int -> ('a, 's) Bap_core_theory.Theory.format Bap_core_theory.Theory.float -> Bap_core_theory.Theory.rmode -> ('a, 's) Bap_core_theory.Theory.format Bap_core_theory.Theory.float

tabulate op ~rank ~size x rmode defines a subset of functions that can be created by approximate, s.t. each value in a table is a concatenation of coefficients (from one with the least rank to one with the most), and each key is first size bits of floating point value.

parameter op

is a name of floating point operation to approximate.

val table : string -> ('r, 's) Bap_core_theory.Theory.format Bap_core_theory.Theory.Float.t Bap_core_theory.Theory.Value.sort -> int -> string

table operation sort rank defines a naming scheme for approximation of rank of an operation for values of sort.

val is_table : Bap_core_theory.Theory.Var.ident -> Base.bool

is_table ident returns true if ident is a table

val operation : Bap_core_theory.Theory.Var.ident -> string

operation ident returns the name of a function, which polynomial coefficients reside in a table referenced by ident. Raise Not_a_table if ident doesn't match to a naming scheme

module Scheme : sig ... end

module contains naming schemes for different math functions